requirements.txt
# Runtime dependencies
pyyaml>=5.1
pyperclip>=1.8.0

tiktoken

